Abstracting out the content wrapper from index.js
